Fixing most of the linting issue raised by golangci-lint
The old linter where run without build flags. This meant that any file
behing a build flag was not linted. This is why we have so many issue
when switching to golangci-lint.

+++ pkg/process/checks/process_nix.go
@@ -21,10 +21,10 @@ func formatUser(fp *process.FilledProcess) *model.ProcessUser {
 		if err == nil {
 			username = u.Username
 		}
-		uid = int32(fp.Uids[0])
+		uid = fp.Uids[0]
 	}
 	if len(fp.Gids) > 0 {
-		gid = int32(fp.Gids[0])
+		gid = fp.Gids[0]
 	}
 
 	return &model.ProcessUser{
